レポート用のコードが完成した

+++ repo3-2-1.cpp
@@ -7,9 +7,9 @@ int main() {
     double d_num1, d_num2, d_ans;
     f_num1 = d_num1 = 20000001;
     f_num2 = d_num2 = 20000000;
-    f_ans = 1/ (sqrt(f_num1) + sqrt(f_num2));
+    f_ans = 1/ (static_cast<float>(sqrt(f_num1)) + static_cast<float>(sqrt(f_num2)));
     d_ans = 1 / (sqrt(d_num1) + sqrt(d_num2));
-    cout<<"Pattern float "<<f_ans<<endl;
-    cout<<"Pattern double "<<d_ans<<endl;
+    cout << "Pattern float " << f_ans << endl;
+    cout << "Pattern double " << d_ans << endl;
     return 0;
 }
